node: Assign remote node identity to all node IPs
Node IPs other than the Cilium internal IPs where only added to the ipcache so
far when encryption was enabled. This commit adds all node IPs to the ipcache
and assign the correct node identity.

The code is cleaned up while doing so. `n.GetNodeIP(false)` always returns the
IPv4 node address and is not dependant on the current address while iterating
over all node addresses. Calling it inside the loop is therefore unnecessary as
the result is always the same. Similarly, the `nodeIP4` variable is redundant
as the nodeIP is always an IPv4 address anyway.

Signed-off-by: Thomas Graf <thomas@cilium.io>